+++ src/xfpm-backlight.c
@@ -72,8 +72,8 @@ struct XfpmBacklightPrivate
     gboolean	    has_hw;
     gboolean	    on_battery;
     
-    gint            last_level;
-    gint 	    max_level;
+    glong            last_level;
+    glong 	    max_level;
     
     gboolean        dimmed;
     gboolean	    block;
@@ -88,7 +88,7 @@ xfpm_backlight_dim_brightness (XfpmBacklight *backlight)
     
     if (xfpm_power_get_mode (backlight->priv->power) == XFPM_POWER_MODE_NORMAL )
     {
-	gint dim_level;
+	glong dim_level;
 	
 	g_object_get (G_OBJECT (backlight->priv->conf),
 		      backlight->priv->on_battery ? BRIGHTNESS_LEVEL_ON_BATTERY : BRIGHTNESS_LEVEL_ON_AC, &dim_level,
@@ -104,7 +104,7 @@ xfpm_backlight_dim_brightness (XfpmBacklight *backlight)
 	
 	dim_level = dim_level * backlight->priv->max_level / 100;
 	
-	XFPM_DEBUG ("Current brightness level before dimming : %i, new %i", backlight->priv->last_level, dim_level);
+	XFPM_DEBUG ("Current brightness level before dimming : %li, new %li", backlight->priv->last_level, dim_level);
 	
 	backlight->priv->dimmed = xfpm_brightness_set_level (backlight->priv->brightness, dim_level);
     }
@@ -268,7 +268,7 @@ xfpm_backlight_reset_cb (EggIdletime *idle, XfpmBacklight *backlight)
     {
 	if ( !backlight->priv->block)
 	{
-	    XFPM_DEBUG ("Alarm reset, setting level to %i", backlight->priv->last_level);
+	    XFPM_DEBUG ("Alarm reset, setting level to %li", backlight->priv->last_level);
 	    xfpm_brightness_set_level (backlight->priv->brightness, backlight->priv->last_level);
 	}
 	backlight->priv->dimmed = FALSE;
@@ -278,7 +278,7 @@ xfpm_backlight_reset_cb (EggIdletime *idle, XfpmBacklight *backlight)
 static void
 xfpm_backlight_button_pressed_cb (XfpmButton *button, XfpmButtonKey type, XfpmBacklight *backlight)
 {
-    gint level;
+    glong level;
     gboolean ret = TRUE;
     
     gboolean enable_brightness, show_popup;
